[simd] Really put each instantiation into its own .cc.
This should hopefully speed up compilation even more.

+++ dune/common/simd/test/standardtest_double.cc
@@ -6,7 +6,7 @@
 namespace Dune {
   namespace Simd {
 
-    template void UnitTest::checkSimdType<unsigned          >();
+    template void UnitTest::checkSimdType<double            >();
 
   } // namespace Simd
 } // namespace Dune
